Execution Redesign
I changed the logic of how tests are executed so that the dynamic part
is loading the test-list but not executing it (the driver behind this
was that I was having issues with the modules, and it was impossible to
see/debug when most of the work was done by the extension. The idea is
that the extension does minimal work (i.e. transform the list of
selected tests into a list of programs that need to be executed) so that
everything is done in the KB itself. In this model, the programs to run
the tests remain and can be executed directly (and manually if one wants
to)... what the extension facilitates is managing test-cases only.  Also
I started translating code to English